+++ CHANGELOG.md
@@ -24,7 +24,9 @@
 ### Other
 
 - GCGI-1313: Refactor reading input_params.json
+- GCGI-1322: Warning instead of error if manually configured sample name does not match provenance
 - GCGI-1323: Support PWGS in mini-Djerba by making summary optional
+
 
 ## v1.5.5: 2024-03-05
 - Further fixes to pWGS Cardea helper to allow multiple projects for one donor